% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{bretagne}
\alias{bretagne}
\title{Communes of Bretagne}
\format{sf}
\source{
Extract from GEOFLA® 2016 v2.2 Communes France Métropolitaine - \url{http://professionnels.ign.fr/geofla}
}
\description{
Delineations of the Bretagne Communes.
}
