% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{.validate_distance_tabs}
\alias{.validate_distance_tabs}
\title{Validate distance dictionaries}
\usage{
.validate_distance_tabs(tabs_list, n_positions)
}
\arguments{
\item{tabs_list}{Matrix dictionary or list of matrix dictionaries.}

\item{n_positions}{Number of barcode positions to use.}
}
\value{
Logical scalar indicating whether \code{tabs_list} is a list.
}
\description{
Internal helper for checking matrix/list dictionary inputs before distance
calculation.
}
