% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemap.R
\name{linemap}
\alias{linemap}
\title{Line Map}
\usage{
linemap(x, var, k = 2, threshold = 1, col = "white", border = "black",
  lwd = 0.5, add = FALSE)
}
\arguments{
\item{x}{a data.frame, two first column must be longitude and latitude of
gridded data}

\item{var}{the name of the variable to plot}

\item{k}{expension factor}

\item{threshold}{threshold of the data to plot}

\item{col}{color for the lines areas}

\item{border}{color for the lines borders}

\item{lwd}{thickness of the lines}

\item{add}{if TRUE add the lines to the current plot}
}
\description{
Plot a line map.
}
\examples{
data("Occitanie")
par(mar=c(0,0,0,0), bg = "ivory2")
plot(sf::st_geometry(regOcc), col="ivory1", border = NA, lwd = 1, lty=3)
mtext(text = "Occitanie, le nouveau Mordor", side = 3, line = -2,
      adj = 0.01, cex = 1.5, col = "ivory4", font = 3)
linemap(x = popOcc, var = "pop", k = 2.5, threshold = 50,
        col = "ivory1", border = "ivory4", lwd = 0.6, add = TRUE)
}
