% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_scaled.R
\name{plot_time_scaled}
\alias{plot_time_scaled}
\title{Plot a time-scaled phylogenetic tree}
\usage{
plot_time_scaled(phy, expand = 1.08)
}
\arguments{
\item{phy}{A phylo object.}

\item{...}{Additional arguments passed to \code{plot.phylo()}.}
}
\value{
Invisibly returns the input tree.
}
\description{
Basic plotting helper for time-scaled phylogenies.
}
