% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{popOcc}
\alias{popOcc}
\title{Popultation Data for Occitanie}
\format{data.frame}
\source{
UMS RIATE
}
\description{
Gridded population of Occitanie
}
\section{Fields}{

\describe{
\item{\code{X}}{longitude}

\item{\code{Y}}{latitude}

\item{\code{pop}}{population}
}}

