% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_subtrees.R
\name{sample_subtrees}
\alias{sample_subtrees}
\title{Sample matched subtrees from two phylogenies}
\usage{
sample_subtrees(phy, phy1, n_keep = 30, seed = NULL)
}
\arguments{
\item{phy}{A phylo object with time MRCA.}

\item{phy1}{A second phylo object with cell division MRCA.}

\item{n_keep}{Integer. Number of tips to retain.}

\item{seed}{Optional integer. Random seed for reproducibility.}
}
\value{
A list containing:
\describe{
\item{keep_tips}{Sampled tip labels}
\item{phy_sub}{Subtree from \code{phy}}
\item{phy_sub1}{Subtree from \code{phy1}}
}
}
\description{
Randomly samples a subset of tip labels from two matched trees and returns
the corresponding subtrees.
}
