% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lineage_tree.R
\name{simulate_lineage_tree}
\alias{simulate_lineage_tree}
\title{Simulate lineage history and construct phylogenetic trees}
\usage{
simulate_lineage_tree(
  Q,
  gRNA_num = 200,
  mutation_rate = seq(0.05, 0.25, 0.05),
  recycle_blocks = 40,
  t_max = 500,
  cell_num = 1,
  which = "sg",
  use_full_edges = TRUE,
  keep_alive_only = TRUE,
  set_all_edge_time_to_one = TRUE,
  rename_cells = TRUE,
  plot = FALSE,
  plot_which = "phy",
  seed = NULL
)
}
\arguments{
\item{Q}{Transition matrix or simulation object.}

\item{gRNA_num}{Integer. Number of gRNAs.}

\item{mutation_rate}{Numeric vector. Mutation rates used for simulation.}

\item{recycle_blocks}{Integer. Number of recycled blocks in preprocessing.}

\item{t_max}{Integer. Maximum simulation time.}

\item{cell_num}{Integer. Initial number of cells.}

\item{which}{Character. guide RNA type, e.g. \code{"sg"} 'hg'.}

\item{use_full_edges}{Logical. Whether to use full edges.}

\item{keep_alive_only}{Logical. Whether to retain only alive cells.}

\item{set_all_edge_time_to_one}{Logical. Whether to set all edge lengths to 1.}

\item{rename_cells}{Logical. Whether to rename cell labels.}

\item{plot}{Logical. Whether to plot the output tree.}

\item{plot_which}{Character. Which tree to plot, e.g. \code{"phy"} tree with exact time  or \code{"phy1"} tree with cell division number.}

\item{seed}{Optional integer. Random seed for reproducibility.}
}
\value{
A list with three components:
\describe{
\item{prep}{Output from \code{prepare_Q_lists()}}
\item{sim}{Output from \code{simulate_lineage()}}
\item{tree}{Output from \code{build_tree_from_sim()}}
}
}
\description{
High-level wrapper for preparing transition matrices, simulating lineage
histories, and building phylogenetic trees.
}
