% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{linemap-package}
\alias{linemap-package}
\title{Line Map Package}
\description{
Create maps made of lines.
}
\note{
These three mains sources gave me the inspiration to create \code{linemap}:\cr
- Joy Division's 'Unknown Pleasures' Cover (\url{https://en.wikipedia.org/wiki/Unknown_Pleasures}) \cr
- the work of James Cheshire (Population Lines: How and Why I Created It - \url{http://spatial.ly/2017/04/population-lines-how-and-why-i-created-it/})\cr
- the work of Ryan Brideau (GeospatialLineGraphs - \url{https://github.com/Brideau/GeospatialLineGraphs})\cr
}
