% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{linemap-package}
\alias{linemap-package}
\title{Line Map Package}
\description{
Create maps made of lines. The package contains two functions: \code{\link{linemap}} and \code{\link{getgrid}}.

\code{\link{linemap}} displays a map made of lines using a data frame of gridded data.

\code{\link{getgrid}} transforms a set of polygons (\code{\link{sf}} objects) into a suitable data frame
for \code{\link{linemap}}.
}
\note{
These three mains sources gave me the inspiration to create \code{linemap}:\cr
- Joy Division's 'Unknown Pleasures' Cover (\url{https://en.wikipedia.org/wiki/Unknown_Pleasures}) \cr
- the work of James Cheshire (Population Lines: How and Why I Created It - \url{http://spatial.ly/2017/04/population-lines-how-and-why-i-created-it/})\cr
- the work of Ryan Brideau (GeospatialLineGraphs - \url{https://github.com/Brideau/GeospatialLineGraphs})\cr
}
